Fix a bunch of pedantic warnings (#8908)
* Fix some ITS related files

* Add MatchITSTPCQC and MatchTPCITS

* Add ClustererTask CookedTracker FastMultEst RecoGeomHelper

* Add Detector

* Change construction order rather than datamember settings

* Remove unused variables

* Fix test, re-implementing Hit addition

* Fix initialisation order in NoisCalibratorSpec

* Fix wrong conversion to int

* Update checkStack.cxx

+++ Detectors/ITSMFT/ITS/tracking/include/ITStracking/Smoother.h
@@ -28,7 +28,7 @@ template <unsigned int D>
 class Smoother
 {
  public:
-  Smoother(TrackITSExt& track, int layer, const ROframe& event, float bZ, o2::base::PropagatorF::MatCorrType corr);
+  Smoother(TrackITSExt& track, size_t layer, const ROframe& event, float bZ, o2::base::PropagatorF::MatCorrType corr);
   ~Smoother();
 
   bool isValidInit() const
@@ -48,7 +48,7 @@ class Smoother
   bool smoothTrack();
 
  private:
-  int mLayerToSmooth;                       // Layer to compute smoothing optimization
+  size_t mLayerToSmooth;                    // Layer to compute smoothing optimization
   float mBz;                                // Magnetic field along Z
   bool mInitStatus;                         // State after the initialization
   o2::base::PropagatorF::MatCorrType mCorr; // Type of correction to use
